ifdef for EXTRA_CHECK, and precalculate time_width_double
git-svn-id: http://ltt.polymtl.ca/svn@733 04897980-b3bd-0310-b5e0-8ef037075253

+++ ltt/branches/poly/ltt/time.h
@@ -90,7 +90,7 @@ static inline double ltt_time_to_double(LttTime t1)
   if(t1.tv_sec > MAX_TV_SEC_TO_DOUBLE)
     g_warning("Precision loss in conversion LttTime to double");
 #endif //EXTRA_CHECK
-  return ((double)t1.tv_sec * NANOSECONDS_PER_SECOND) + (double)t1.tv_nsec;
+  return ((double)t1.tv_sec * (double)NANOSECONDS_PER_SECOND) + (double)t1.tv_nsec;
 }
